+++ backend/src/main/java/com/example/iis/dto/PlantOrderRecommendationResponse.java
@@ -1,12 +1,11 @@
 package com.example.iis.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record PlantOrderRecommendationResponse(
-        BigDecimal budget,
-        BigDecimal total,
-        BigDecimal remainingBudget,
+        Integer budget,
+        Integer total,
+        Integer remainingBudget,
         Integer itemCount,
         List<PlantOrderRecommendationItemResponse> items
 ) {
